+++ nf_core/__main__.py
@@ -808,7 +808,12 @@ def create_test_yml(ctx, tool, run_tests, output, force, no_prompts):
 @click.argument("tool", type=str, required=False, metavar="<tool> or <tool/subtool>")
 @click.option("-d", "--dir", type=click.Path(exists=True), default=".", metavar="<pipeline/modules directory>")
 @click.option(
-    "-r", "--registry", type=str, metavar="<registry>", default="quay.io", help="Registry to use for containers"
+    "-r",
+    "--registry",
+    type=str,
+    metavar="<registry>",
+    default=None,
+    help="Registry to use for containers. If not specified will interpret it from nextflow.config file at attribute docker.registry",
 )
 @click.option("-k", "--key", type=str, metavar="<test>", multiple=True, help="Run only these lint tests")
 @click.option("-a", "--all", is_flag=True, help="Run on all modules")
@@ -842,6 +847,7 @@ def lint(
         module_lint = ModuleLint(
             dir,
             fail_warned=fail_warned,
+            registry=ctx.params["registry"],
             remote_url=ctx.obj["modules_repo_url"],
             branch=ctx.obj["modules_repo_branch"],
             no_pull=ctx.obj["modules_repo_no_pull"],
@@ -849,7 +855,6 @@ def lint(
         )
         module_lint.lint(
             module=tool,
-            registry=registry,
             key=key,
             all_modules=all,
             print_results=True,